# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin Fisiologico API",
    "description": "Stateless physiology analytics API. HTTP routers are thin; application services orchestrate engines under `engines/`.",
    "version": "5.1.0",
    "x-codegen": {
      "typescript_output": "frontend/src/api/generated/schema.ts",
      "regenerate": "make openapi-frontend"
    }
  },
  "paths": {
...
    "/performance/neuromuscular-profile": {
      "post": {
        "tags": [
          "performance"
        ],
        "summary": "Neuromuscular sprint profile",
        "description": "Pmax, cadence-at-peak, repeat-sprint profile from FIT or power stream.",
        "operationId": "performanceNeuromuscularProfile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_performanceNeuromuscularProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/power-source/normalize": {
      "post": {
        "tags": [
          "performance"
        ],
        "summary": "Detect power source offsets",
        "description": "Compare indoor trainer vs outdoor PM systematic offsets.",
        "operationId": "powerSourceNormalize",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PowerSourceNormalizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_performanceNeuromuscularProfile": {
        "properties": {
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg",
            "default": 70.0
          },
          "sprint_threshold_w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Threshold W"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "power_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Json"
          }
        },
        "type": "object",
        "title": "Body_performanceNeuromuscularProfile"
      },
...
      "EnginePayload": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "EnginePayload",
        "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PowerSourceNormalizationRequest": {
        "properties": {
          "activities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Activities",
            "description": "Activities with source_id and MMP signatures."
          },
          "baseline_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Source Id"
          },
          "warning_threshold_pct": {
            "type": "number",
            "title": "Warning Threshold Pct",
            "default": 3.0
          },
          "severe_threshold_pct": {
            "type": "number",
            "title": "Severe Threshold Pct",
            "default": 6.0
          }
        },
        "type": "object",
        "title": "PowerSourceNormalizationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Configurable deployment",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000",
          "description": "Set via VITE_API_BASE_URL in the frontend"
        }
      }
    }
  ],
...
}